+++ components/organisms/PhaseBanner.stories.js
@@ -9,8 +9,15 @@ export default {
 const Template = (args) => <PhaseBanner {...args} />;
 
 export const Primary = Template.bind({});
+export const WithFeedback = Template.bind({});
 
 Primary.args = {
   phase: "PhaseBanner Phase",
   children: "PhaseBanner Text",
+};
+
+WithFeedback.args = {
+  phase: "PhaseBanner Phase",
+  children: "PhaseBanner Text",
+  feedbackActive: true,
 };